Mod:refactor to add base_registry
